Issue #1394: Unplug BrowserNavigationOverlay from WebRenderFrag
