+++ app/src/androidTest/java/org/mozilla/tv/firefox/ui/robots/NavigationOverlayRobot.kt
@@ -197,6 +197,6 @@ private fun turboButton() = onView(withId(R.id.turboButton))
 private fun settingsButton() = onView(withId(R.id.navButtonSettings))
 private fun urlBar() = onView(withId(R.id.navUrlInput))
 private fun homeTiles() = onView(withId(R.id.tileContainer))
-private fun overlay() = onView(withId(R.id.browserOverlay))
+private fun overlay() = onView(withId(R.layout.browser_overlay))
 private fun desktopModeButton() = onView(withId(R.id.desktopModeButton))
 private fun pocketMegaTile() = onView(withId(R.id.pocketVideosContainer))
