+++ app/src/main/res/layout/fragment_browser.xml
@@ -58,11 +58,4 @@
         android:focusable="false"
         android:descendantFocusability="blocksDescendants"
         />
-
-    <org.mozilla.tv.firefox.navigationoverlay.BrowserNavigationOverlay
-        android:id="@+id/browserOverlay"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:visibility="gone"
-        />
 </FrameLayout>
